news_crawler/requirements.txt
feedparser==5.2.1
mysql-connector-python-rf==2.2.2